refactor: http-client 및 useOverlay 경로 수정, 새로운 훅 및 유틸리티 추가
