+++ src/shared/lib/highlightText.tsx
